refactor(tests): share typed run_async helper
Deduplicate the _run_async helpers that existed in both
test_dependencies and test_images into tests.async_helpers.run_async.
