+++ tests/utils/test_images.py
@@ -1,9 +1,8 @@
 import pytest
 from PIL import Image
 import io
-import asyncio
-from concurrent.futures import ThreadPoolExecutor
 from unittest.mock import AsyncMock, MagicMock
+from tests.async_helpers import run_async
 from utils.core.images import (
     validate_and_process_image,
     read_upload_with_size_limit,
@@ -22,12 +21,6 @@ def create_test_image(width: int, height: int, format: str = "PNG") -> bytes:
     output = io.BytesIO()
     image.save(output, format=format)
     return output.getvalue()
-
-
-def _run_async(coro):
-    """Run a coroutine when pytest may already have an event loop active."""
-    with ThreadPoolExecutor(max_workers=1) as executor:
-        return executor.submit(asyncio.run, coro).result()
 
 
 def test_webp_dependencies_are_installed():
@@ -109,7 +102,7 @@ def test_read_upload_with_size_limit_rejects_oversized_stream():
     )
 
     with pytest.raises(InvalidImageError) as exc_info:
-        _run_async(read_upload_with_size_limit(upload, MAX_FILE_SIZE))
+        run_async(read_upload_with_size_limit(upload, MAX_FILE_SIZE))
 
     assert "File too large" in str(exc_info.value.detail)
     assert upload.read.await_count == 2
@@ -119,7 +112,7 @@ def test_read_upload_with_size_limit_accepts_valid_stream():
     upload = MagicMock()
     upload.read = AsyncMock(side_effect=[b"abc", b"def", b""])
 
-    data = _run_async(read_upload_with_size_limit(upload, MAX_FILE_SIZE))
+    data = run_async(read_upload_with_size_limit(upload, MAX_FILE_SIZE))
 
     assert data == b"abcdef"
